Moved files to same dir as ScrumBoardPage
